Build documentation with pandoc, graphviz & gnuplot
I've been using a couple of scripts locally to work on Markdown
documentation in a semi-WYSIWYG fashion (you edit in one window while
formatted updates appear in another).

This script will start an HTTP server and a process to rebuild all
Markdown, Graphviz and GNUPlot files whenever a file changes in that
directory.

Publishing this as a docker image to make it easier to get access to
this when working on other machines. It should be considered less than
alpha quality code. There are a number of improvements which I'd like to
make. Not least of which to change how the output files are stored and
to use a proper build system like `make`.